optimize centACL.php
git-svn-id: http://svn.centreon.com/trunk/centreon@7074 6bcd3966-0018-0410-8128-fd23d134de7e

+++ cron/centAcl.php
@@ -24,6 +24,15 @@
 	include_once $centreon_path."/www/include/common/common-Func.php";
 	include_once $centreon_path."/www/include/common/common-Func-ACL.php";
 	
+	/*
+	 * Init values
+	 */
+	
+	$debug = 0;
+	
+	/*
+	 * Init functions
+	 */
 	function microtime_float2() 	{
 	   list($usec, $sec) = explode(" ", microtime());
 	   return ((float)$usec + (float)$sec);
@@ -35,8 +44,8 @@
 									"WHERE acl_groups.acl_group_id = acl_res_group_relations.acl_group_id " .
 									"AND acl_res_group_relations.acl_res_id = acl_resources.acl_res_id " .
 									"AND `acl_resources`.`acl_res_activate` = '1' " .
-									"AND (acl_groups.acl_group_activate = '1' ".			
-									"OR acl_resources.changed = '1')");
+									"AND acl_groups.acl_group_activate = '1' ".			
+									"AND acl_resources.changed = '1'");
 	while ($result =& $DBRESULT1->fetchRow())
 		$tabGroups[$result["acl_group_id"]] = 1;
 	
@@ -44,15 +53,19 @@
 	 * Purge datas
 	 */
 	$strBegin = "INSERT INTO `centreon_acl` ( `host_name` , `service_description` , `group_id` ) VALUES ";
-	
+
+	$cpt = 0;	
 	foreach ($tabGroups as $acl_group_id => $acl_res_id){
 		$tabElem = array();
-		/*
-		 * Delete old datas for this groups
-		 */
-		$DBRESULT =& $pearDBndo->query("DELETE FROM `centreon_acl` WHERE `group_id` = '".$acl_group_id."'");
-		if (PEAR::isError($DBRESULT))
-			print "DB Error : ".$DBRESULT->getDebugInfo()."<br />";
+		
+		if ($cpt == 0) {
+			/*
+			 * Delete old datas for this groups
+			 */
+			$DBRESULT =& $pearDBndo->query("DELETE FROM `centreon_acl` WHERE `group_id` = '".$acl_group_id."'");
+			if (PEAR::isError($DBRESULT))
+				print "DB Error : ".$DBRESULT->getDebugInfo()."<br />";			
+		}
 		
 		/*
 		 * Select 
@@ -60,7 +73,8 @@
 		
 		$Host = array();
 		$DBRESULT2 =& $pearDB->query("SELECT `acl_res_id` FROM `acl_res_group_relations` WHERE `acl_group_id` = '".$acl_group_id."'");			
-		//$time_start = microtime_float2();
+		if ($debug)
+			$time_start = microtime_float2();
 		while ($res2 =& $DBRESULT2->fetchRow()){
 	
 			/* ------------------------------------------------------------------ */
@@ -69,7 +83,7 @@
 			 * Get all Hosts 
 			 */
 			$DBRESULT3 =& $pearDB->query("SELECT host_id, host_name FROM `host`, `acl_resources_host_relations` WHERE acl_res_id = '".$res2["acl_res_id"]."' AND acl_resources_host_relations.host_host_id = host.host_id AND host.host_register = '1' AND host.host_activate = '1'");
-		  	while ($h = $DBRESULT3->fetchRow())
+		  	while ($h =& $DBRESULT3->fetchRow())
 				$Host[$h["host_id"]] = $h["host_name"];
 			$DBRESULT3->free();
 				
@@ -77,7 +91,7 @@
 		  	 * Get all host in hostgroups
 		  	 */
 			$DBRESULT3 =& $pearDB->query("SELECT `hg_id` FROM `hostgroup`, `acl_resources_hg_relations` WHERE acl_res_id = '".$res2["acl_res_id"]."' AND acl_resources_hg_relations.hg_hg_id = hostgroup.hg_id");
-	  		while ($hostgroup = $DBRESULT3->fetchRow()){
+	  		while ($hostgroup =& $DBRESULT3->fetchRow()){
 				$DBRESULT4 =& $pearDB->query("SELECT host_host_id, host_name FROM `hostgroup_relation`, `host` WHERE hostgroup_hg_id = '".$hostgroup["hg_id"]."' AND host.host_id = hostgroup_relation.host_host_id");
 	  			while ($host_hostgroup = $DBRESULT4->fetchRow())
 					$Host[$host_hostgroup["host_host_id"]] = $host_hostgroup["host_name"];
@@ -94,8 +108,7 @@
 						unset($Host[$h["host_id"]]);
 			$DBRESULT3->free();
 			
-			$str = "";
-	
+			$str = "";	
 			foreach ($Host as $key => $value){
 				$tab = getAuthorizedServicesHost($key, $acl_group_id, $res2["acl_res_id"]);
 				foreach ($tab as $desc => $id){
@@ -103,6 +116,7 @@
 						$tabElem[$value] = array();
 					$tabElem[$value][$desc] = 1;
 				}	 
+				unset($tab);
 			}
 
 			/*
@@ -113,19 +127,24 @@
 												"AND host.host_id = servicegroup_relation.host_host_id " .
 												"AND service.service_id = servicegroup_relation.service_service_id " .
 												"AND servicegroup_relation.servicegroup_sg_id = acl_resources_sg_relations.sg_id");
-			if ($DBRESULT3->numRows())
-		  		while ($h = $DBRESULT3->fetchRow()){
+			if ($DBRESULT3->numRows()) {
+		  		while ($h =& $DBRESULT3->fetchRow()){
 					if (!isset($tabElem[$h["host_id"]]))
 						$tabElem[$h["host_id"]] = array();
 		  			$tabElem[$h["host_name"]][$h["service_description"]] = 1;
 		  		}
+			}
 			$DBRESULT3->free();
 			
 			/* ------------------------------------------------------------------ */
 		}
-		//$time_end = microtime_float2(); $now = $time_end - $time_start; print round($now,3) . _(" seconds \n");
-	
 		$DBRESULT2->free();
+
+		if ($debug) {
+			$time_end = microtime_float2(); 
+			$now = $time_end - $time_start; 
+			print round($now,3) . _(" seconds \n");
+		}
 		
 		if (count($tabElem)){
 			foreach ($tabElem as $host => $svc_list){
@@ -136,13 +155,14 @@
 				}
 			}
 			$DBRESULTNDO =& $pearDBndo->query($strBegin.$str);
-			if (PEAR::isError($DBRESULTNDO))
+			if (PEAR::isError($DBRESULTNDO)) {
 				print "DB Error : ".$DBRESULTNDO->getDebugInfo()."<br />";
-			else {
+			} else {
 				$DBRESULT3 =& $pearDB->query("UPDATE `acl_resources` SET `changed` = '0'");
 				if (PEAR::isError($DBRESULT3))
 					print "DB Error : ".$DBRESULT3->getDebugInfo()."<br />";
 			}
 		}	
+		$cpt++;
 	}
 ?>
